fix: address PR review issues — extract shared code, update validation, fix redundant logic
- Extract unit classification (classifyUnit, keyword arrays, UnitClass,
  UNIT_CLASS_COLORS) from OverviewTab and MilitaryTab into helpers.ts
- Add game_religions and player_wonders to Worker schema validation as
  required fields with bounds checking
- Add missing player_wonders assertion in share.rs test
- Remove redundant >= 10000 branch in formatValue (identical to >= 1000)
- Remove unnecessary ?? [] fallbacks in api-web.ts
