+++ src/lib/game-detail/MilitaryTab.svelte
@@ -9,7 +9,7 @@
 	import { Select } from "bits-ui";
 	import { formatEnum } from "$lib/utils/formatting";
 	import { CHART_THEME } from "$lib/config";
-	import { type TableState, getPlayerColor, toggleSort } from "./helpers";
+	import { type TableState, type UnitClass, getPlayerColor, toggleSort, classifyUnit, UNIT_CLASS_COLORS } from "./helpers";
 
 	let {
 		gameDetails,
@@ -64,32 +64,6 @@
 	);
 
 	// ─── Army composition pie charts ─────────────────────────────────
-	const RANGED_KEYWORDS = ["ARCHER", "BOWMAN", "CROSSBOW", "SLINGER", "SKIRMISHER", "JAVELINEER", "CLUB_THROWER"];
-	const CAVALRY_KEYWORDS = ["CAVALRY", "CHARIOT", "CATAPHRACT", "HORSEMAN", "RIDER", "LANCER", "MOUNTED", "ELEPHANT", "CAMEL"];
-	const SIEGE_KEYWORDS = ["BALLISTA", "ONAGER", "SIEGE", "CATAPULT", "RAM", "TORSION"];
-	const NAVAL_KEYWORDS = ["BIREME", "TRIREME", "DROMON", "QUINQUEREME", "PENTECONTER", "GALLEY", "SHIP", "LONGBOAT"];
-	const SUPPORT_KEYWORDS = ["WORKER", "MILITIA", "SETTLER", "SCOUT", "CARAVAN", "DISCIPLE", "DIPLOMAT", "MISSIONARY", "SPY", "MONK", "CLERIC", "PROPHET", "ENVOY", "TRADER"];
-
-	type UnitClass = "Infantry" | "Ranged" | "Cavalry" | "Siege" | "Naval";
-
-	function classifyUnit(unitType: string): UnitClass | null {
-		const upper = unitType.toUpperCase();
-		if (SUPPORT_KEYWORDS.some((k) => upper.includes(k))) return null;
-		if (NAVAL_KEYWORDS.some((k) => upper.includes(k))) return "Naval";
-		if (SIEGE_KEYWORDS.some((k) => upper.includes(k))) return "Siege";
-		if (CAVALRY_KEYWORDS.some((k) => upper.includes(k))) return "Cavalry";
-		if (RANGED_KEYWORDS.some((k) => upper.includes(k))) return "Ranged";
-		return "Infantry";
-	}
-
-	const UNIT_CLASS_COLORS: Record<UnitClass, string> = {
-		Infantry: "#C87941",
-		Ranged: "#B8860B",
-		Cavalry: "#CD853F",
-		Siege: "#A0522D",
-		Naval: "#8B4513",
-	};
-
 	type ArmyPieData = {
 		nation: string | null;
 		nationLabel: string;
